Move goals from AttackingAnimatedEntity out as their own classes
