+++ src/main/java/com/mraof/minestuck/entity/ai/MoveToTargetGoal.java
@@ -1,0 +1,21 @@
+package com.mraof.minestuck.entity.ai;
+
+import net.minecraft.entity.CreatureEntity;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.ai.goal.MeleeAttackGoal;
+
+/**
+ * The same as MeleeAttackGoal, except that the goal does not handle the actual attack.
+ */
+public class MoveToTargetGoal extends MeleeAttackGoal
+{
+	public MoveToTargetGoal(CreatureEntity entity, float speed, boolean followsUnseenTarget)
+	{
+		super(entity, speed, followsUnseenTarget);
+	}
+	
+	@Override
+	protected void checkAndPerformAttack(LivingEntity enemy, double distToEnemySqr)
+	{
+	}
+}
