+++ src/main/java/com/mraof/minestuck/entity/underling/BasiliskEntity.java
@@ -1,5 +1,7 @@
 package com.mraof.minestuck.entity.underling;
 
+import com.mraof.minestuck.entity.ai.MoveToTargetGoal;
+import com.mraof.minestuck.entity.ai.SlowAttackWhenInRangeGoal;
 import com.mraof.minestuck.item.crafting.alchemy.GristHelper;
 import com.mraof.minestuck.item.crafting.alchemy.GristSet;
 import com.mraof.minestuck.item.crafting.alchemy.GristType;
@@ -53,7 +55,7 @@ public class BasiliskEntity extends UnderlingEntity implements IAnimatable
 	protected void registerGoals()
 	{
 		super.registerGoals();
-		this.goalSelector.addGoal(2, new SlowAttackInPlaceGoal(this, 4, 10));
+		this.goalSelector.addGoal(2, new SlowAttackWhenInRangeGoal.InPlace(this, 4, 10));
 		this.goalSelector.addGoal(3, new MoveToTargetGoal(this, 1F, false));
 	}
